requirements.txt
claude-agent-sdk>=0.1.0
mcp>=1.6.0
python-telegram-bot>=21.0
aiohttp>=3.9
apscheduler>=3.10
faster-whisper>=1.0
pypdf>=4.0
duckduckgo-search>=6.0
beautifulsoup4>=4.12
wikipedia>=1.4
youtube-transcript-api>=0.6
arxiv>=2.1
feedparser>=6.0
qrcode>=7.4
Pillow>=10.0
icalendar>=5.0
caldav>=1.3
pytest>=8.0
pytest-asyncio>=0.23
python-dotenv>=1.0
pyyaml>=6.0
telethon>=1.36
